requirements.txt
#statsmodels==0.10.1
numpy==1.17.2
statsmodels==0.9.0
pandas==0.24.2
#grpcio==1.31.0
grpcio
matplotlib==3.0.3
cassandra_driver==3.24.0
grpcio-tools
#grpc==1.0.0
#grpc
protobuf==3.13.0
python_dateutil==2.8.1
